привел некоторые участки кода таблицы в порядок
