+++ src/0-shared/components/TableSortButton/TableSortButton.scss
@@ -1,5 +1,6 @@
 .TableSortButton {
     opacity: 10%;
+    padding: 5px;
 
     &--active {
         opacity: 100%;
